Moved the function get_pypi_latest_version() from util.py to apio_upgrade.py. No functional change.
