+++ apio/utils/util.py
@@ -23,7 +23,6 @@ from threading import Thread
 from pathlib import Path
 from varname import argname
 from serial.tools.list_ports import comports
-import requests
 from apio.utils import env_options
 from apio.common.apio_console import cout, cerror
 
@@ -203,51 +202,6 @@ def exec_command(*args, **kwargs) -> CommandResult:
     # -- All done.
     result = CommandResult(out_text, err_text, exit_code)
     return result
-
-
-def get_pypi_latest_version() -> str:
-    """Get the latest stable version of apio from Pypi
-    Internet connection is required
-    Returns: A string with the version (Ex: "0.9.0")
-    Exits on an error.
-    """
-
-    # -- Read the latest apio version from pypi
-    # -- More information: https://warehouse.pypa.io/api-reference/json.html
-    try:
-        req = requests.get(
-            "https://pypi.python.org/pypi/apio/json", timeout=10
-        )
-        req.raise_for_status()
-
-    # -- Connection error
-    except requests.exceptions.ConnectionError as e:
-        cout(str(e), style="yellow")
-        cerror("Connection error while accessing Pypi.")
-        sys.exit(1)
-
-    # -- HTTP Error
-    except requests.exceptions.HTTPError as e:
-        cout(str(e), style="yellow")
-        cerror("HTTP error while accessing Pypi.")
-        sys.exit(1)
-
-    # -- Timeout!
-    except requests.exceptions.Timeout as e:
-        cout(str(e), style="yellow")
-        cerror("HTTP timeout while accessing Pypi.")
-        sys.exit(1)
-
-    # -- Another error
-    except requests.exceptions.RequestException as e:
-        cout(str(e), style="yellow")
-        cerror("HTTP exception while accessing Pypi.")
-        sys.exit(1)
-
-    # -- Get the version field from the json response
-    version = req.json()["info"]["version"]
-
-    return version
 
 
 def resolve_project_dir(
